requirements.txt
django==4.1.7
django-ckeditor==6.5.1
requests==2.28.1
ptvsd==4.3.2 # Required for debugging.
google-cloud-secret-manager==2.15.1
django-environ==0.9.0
psycopg2-binary==2.9.5
django-storages==1.13.2
google-cloud-core==2.3.2
GoogleAppEngineCloudStorageClient==1.9.15.0
google-cloud-storage==2.7.0
